% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castorapiclient.R
\name{CastorApiClient}
\alias{CastorApiClient}
\title{API client for interacting with Castor EDC API}
\description{
API client for interacting with Castor EDC API
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_token_url}}{API token URL}

\item{\code{api_base_url}}{Base URL}

\item{\code{token}}{Access token}

\item{\code{studies}}{Study list}

\item{\code{nr_retries}}{Nr. of retries}

\item{\code{retry_waiting_time}}{Time in seconds to wait before retrying}

\item{\code{data}}{Study data records}

\item{\code{field_defs}}{Study field definitions}

\item{\code{optiongroups}}{Study option groups}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CastorApiClient-new}{\code{CastorApiClient$new()}}
\item \href{#method-CastorApiClient-connect}{\code{CastorApiClient$connect()}}
\item \href{#method-CastorApiClient-is_connected}{\code{CastorApiClient$is_connected()}}
\item \href{#method-CastorApiClient-get_studies}{\code{CastorApiClient$get_studies()}}
\item \href{#method-CastorApiClient-get_study_name_by_id}{\code{CastorApiClient$get_study_name_by_id()}}
\item \href{#method-CastorApiClient-get_study_id_by_name}{\code{CastorApiClient$get_study_id_by_name()}}
\item \href{#method-CastorApiClient-get_study_data_as_csv}{\code{CastorApiClient$get_study_data_as_csv()}}
\item \href{#method-CastorApiClient-load_csv_data}{\code{CastorApiClient$load_csv_data()}}
\item \href{#method-CastorApiClient-get_field_type}{\code{CastorApiClient$get_field_type()}}
\item \href{#method-CastorApiClient-set_dataframe_data_types}{\code{CastorApiClient$set_dataframe_data_types()}}
\item \href{#method-CastorApiClient-get_study_data}{\code{CastorApiClient$get_study_data()}}
\item \href{#method-CastorApiClient-save_records}{\code{CastorApiClient$save_records()}}
\item \href{#method-CastorApiClient-save_field_defs}{\code{CastorApiClient$save_field_defs()}}
\item \href{#method-CastorApiClient-save_na_counts}{\code{CastorApiClient$save_na_counts()}}
\item \href{#method-CastorApiClient-clone}{\code{CastorApiClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-new}{}}}
\subsection{Method \code{new()}}{
Initializes API client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$new(
  client_id,
  client_secret,
  nr_retries = 5,
  retry_waiting_time = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client_id}}{Client ID}

\item{\code{client_secret}}{Client secret}

\item{\code{nr_retries}}{Nr. of retries}

\item{\code{retry_waiting_time}}{Time in seconds to wait}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-connect"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-connect}{}}}
\subsection{Method \code{connect()}}{
Connects to API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$connect(client_id, client_secret, token_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client_id}}{Client ID}

\item{\code{client_secret}}{Client secret}

\item{\code{token_url}}{URL for retrieving access token}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Token string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-is_connected"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-is_connected}{}}}
\subsection{Method \code{is_connected()}}{
Checks whether connected or not
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$is_connected()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if connected, \code{FALSE} if not
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-get_studies"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-get_studies}{}}}
\subsection{Method \code{get_studies()}}{
Gets list of studies
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$get_studies()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Study list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-get_study_name_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-get_study_name_by_id}{}}}
\subsection{Method \code{get_study_name_by_id()}}{
Gets study name for given ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$get_study_name_by_id(study_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_id}}{Study ID}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Study name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-get_study_id_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-get_study_id_by_name}{}}}
\subsection{Method \code{get_study_id_by_name()}}{
Gets study ID for given name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$get_study_id_by_name(study_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_name}}{Study name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Study ID
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-get_study_data_as_csv"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-get_study_data_as_csv}{}}}
\subsection{Method \code{get_study_data_as_csv()}}{
Gets study data of given data type in CSV format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$get_study_data_as_csv(study_id, data_type, tmp_dir = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_id}}{Study ID}

\item{\code{data_type}}{Data type to retrieve. Must be: "data", "structure" or "optiongroups"}

\item{\code{tmp_dir}}{Temporary directory to save data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Study data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-load_csv_data"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-load_csv_data}{}}}
\subsection{Method \code{load_csv_data()}}{
Loads CSV as dataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$load_csv_data(csv_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{csv_data}}{CSV data in text format}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-get_field_type"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-get_field_type}{}}}
\subsection{Method \code{get_field_type()}}{
Returns field type for given field name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$get_field_type(field_name, field_defs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field_name}}{Field name}

\item{\code{field_defs}}{Dataframe with field definitions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Field type as string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-set_dataframe_data_types"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-set_dataframe_data_types}{}}}
\subsection{Method \code{set_dataframe_data_types()}}{
Sets dataframe column types based on field definitions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$set_dataframe_data_types(df, field_defs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{Dataframe}

\item{\code{field_defs}}{Dataframe with field definitions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-get_study_data"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-get_study_data}{}}}
\subsection{Method \code{get_study_data()}}{
Returns full study data for given study name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$get_study_data(study_name, tmp_dir = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_name}}{Study name}

\item{\code{tmp_dir}}{Temporary directory to save study data to file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named list of records, field definitions and option groups
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-save_records"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-save_records}{}}}
\subsection{Method \code{save_records()}}{
Saves study records to given file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$save_records(
  file_path = file.path(path.expand("~"), "records.Rdata")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_path}}{Path to data file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-save_field_defs"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-save_field_defs}{}}}
\subsection{Method \code{save_field_defs()}}{
Saves study field definitions to given file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$save_field_defs(
  file_path = file.path(path.expand("~"), "field_defs.Rdata")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_path}}{Path to data file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-save_na_counts"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-save_na_counts}{}}}
\subsection{Method \code{save_na_counts()}}{
Saves NA counts to file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$save_na_counts(
  file_path = file.path(path.expand("~"), "na_counts.Rdata")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_path}}{Path to data file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
