% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{CastorApiCredentials}
\alias{CastorApiCredentials}
\title{Class for handling Castor API credentials}
\description{
Class for handling Castor API credentials

Class for handling Castor API credentials
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client_id_file_name}}{Client ID file name (default: "castorclientid.txt")}

\item{\code{client_id_file_path}}{Full client ID file path}

\item{\code{client_id}}{Client ID}

\item{\code{client_secret_file_name}}{Client secret file name (default: "castorclientsecret.txt")}

\item{\code{client_secret_file_path}}{Full client secret file path}

\item{\code{client_secret}}{Client secret}

\item{\code{study_name_file_name}}{Study name file name (default: castorstudy.txt")}

\item{\code{study_name_file_path}}{Study name file path}

\item{\code{study_name}}{Study name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CastorApiCredentials-new}{\code{CastorApiCredentials$new()}}
\item \href{#method-CastorApiCredentials-load_client_id}{\code{CastorApiCredentials$load_client_id()}}
\item \href{#method-CastorApiCredentials-save_client_id}{\code{CastorApiCredentials$save_client_id()}}
\item \href{#method-CastorApiCredentials-load_client_secret}{\code{CastorApiCredentials$load_client_secret()}}
\item \href{#method-CastorApiCredentials-save_client_secret}{\code{CastorApiCredentials$save_client_secret()}}
\item \href{#method-CastorApiCredentials-load_study_name}{\code{CastorApiCredentials$load_study_name()}}
\item \href{#method-CastorApiCredentials-save_study_name}{\code{CastorApiCredentials$save_study_name()}}
\item \href{#method-CastorApiCredentials-clone}{\code{CastorApiCredentials$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiCredentials-new"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiCredentials-new}{}}}
\subsection{Method \code{new()}}{
Initializes Castor API credentials handler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiCredentials$new(
  client_id_file_name = NULL,
  client_secret_file_name = NULL,
  study_name_file_name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client_id_file_name}}{Client ID file name}

\item{\code{client_secret_file_name}}{Client secret file name}

\item{\code{study_name_file_name}}{Study name file name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiCredentials-load_client_id"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiCredentials-load_client_id}{}}}
\subsection{Method \code{load_client_id()}}{
Loads client ID from file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiCredentials$load_client_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Client ID or NULL if not found
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiCredentials-save_client_id"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiCredentials-save_client_id}{}}}
\subsection{Method \code{save_client_id()}}{
Saves client ID to file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiCredentials$save_client_id(client_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client_id}}{Client ID}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiCredentials-load_client_secret"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiCredentials-load_client_secret}{}}}
\subsection{Method \code{load_client_secret()}}{
Loads client secret from file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiCredentials$load_client_secret()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Client secret or NULL if not found
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiCredentials-save_client_secret"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiCredentials-save_client_secret}{}}}
\subsection{Method \code{save_client_secret()}}{
Saves client secret to file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiCredentials$save_client_secret(client_secret)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client_secret}}{Client secret}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiCredentials-load_study_name"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiCredentials-load_study_name}{}}}
\subsection{Method \code{load_study_name()}}{
Loads study name from file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiCredentials$load_study_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Study name or NULL if not found
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiCredentials-save_study_name"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiCredentials-save_study_name}{}}}
\subsection{Method \code{save_study_name()}}{
Saves study name to file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiCredentials$save_study_name(study_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_name}}{Study name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CastorApiCredentials-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CastorApiCredentials-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CastorApiCredentials$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
