% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAchillesMonthlyData.r
\name{getAchillesMonthlyData}
\alias{getAchillesMonthlyData}
\title{Get Achilles monthly data by concept_id or set of concept_ids}
\usage{
AchillesResultSet <- getAchillesMonthlyData (connectionDetails, cdmSchema, resultsSchema, conceptId)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSchema}{A fully qualified CDM schema name.}

\item{resultsSchema}{A fully qualified results schema name.}

\item{conceptId}{One more more valid concept_ids.}
}
\value{
A dataframe of query results.
}
\description{
Return monthly count, prevalence, and proportion within year data from ACHILLES_RESULTS
for a specified concept_id or set of concept_ids.
}
\details{
\preformatted{Currently supported Achilles monthly analyses are:
202  - Visit Occurrence
402  - Condition occurrence
602  - Procedire Occurrence
702  - Drug Exposure
802  - Observation
1802 - Measurement
2102 - Device}
}
