% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAchillesTs.r
\name{getAchillesTs}
\alias{getAchillesTs}
\title{Return a monthly multivariate time series of Achilles data from a data frame or concept_ids}
\usage{
# Method 1: Create the time series in two steps then access the individual time series.
      # First, get results from Achilles for one or more concept_ids, then transform the results into a time series object.
        AchillesData <- getAchillesMonthlyData (connectionDetails, cdmSchema, resultsSchema, conceptId) 
        achillesTs   <- getAchillesTs(resultSet = AchillesData)
        prevTs       <- achillesTs[,"PREVALENCE"]
        countTs      <- achillesTs[,"COUNT_VALUE"]
        pwyTs        <- achillesTs[,"PROPORTION_WITHIN_YEAR"]
      # Method 2: Create the time series in a single step
        achillesTs   <- getAchillesTs(connectionDetails, cdmSchema, resultsSchema, conceptId)
        prevTs       <- achillesTs[,"PREVALENCE"]
        countTs      <- achillesTs[,"COUNT_VALUE"]
        pwyTs        <- achillesTs[,"PROPORTION_WITHIN_YEAR"]
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmSchema}{A fully qualified CDM schema name.}

\item{resultsSchema}{A fully qualified results schema name.}

\item{conceptId}{One more more valid concept_ids.}

\item{resultSet}{A dataframe in the proper format (e.g., one returned by \cr\code{getAchillesMonthlyData})}
}
\value{
A multivariate time series object
}
\description{
Build a multivariate time series object from one or more concept_ids or a dataframe in the proper format.
If supplying a data frame from which to create the time series, it must have the following columns:
START_DATE, COUNT_VALUE, PREVALENCE, and PROPORTION_WITHIN_YEAR.
}
\details{
As the name suggests, this function was designed with Achilles in mind.  An Achilles time series is a multivariate
time series object consisting of three types of time series: prevalence, count, and proportion within year.  The object
returned by this function should not be used in your analyses.  Instead, you should access the individual time series
as per the exaples above.
NB: If you supply one or more concept_ids, the data are retrieved from Achilles analysis ids 202,402,602,702,802,1802,
and 2102 (corresponding to your concept ids).
If you supply a data frame, then it need not come from Achilles as per the example above.  It could be any data
frame in the correct format.
}
