% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountTs.r
\name{getCountTs}
\alias{getCountTs}
\title{Build a univariate time series of counts (or distinct counts) from a specified table and columns}
\usage{

}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{dbSchema}{A fully qualified CDM or Results schema name.}

\item{tableName}{A  valid table name.}

\item{dateColumnName}{A valid date column in the specified table.}

\item{startDate}{A string date in the "YYYY-MM-DD" format}

\item{endDate}{A string date in the "YYYY-MM-DD" format}

\item{dataColumnName}{The column in the specified table to count.}

\item{isCountDistinct}{A boolean indicating whether or not to count distinct rows of the specified data column.}

\item{frequency}{A string indicating the number of observations per unit of time in the time series.
Acceptable values are "day","week","month","quarter", and "year".}
}
\value{
A univariate time series object
}
\description{
Build a univariate time series object from a specified table within a given data range.
}
