% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeighborhoodDf.r
\name{getNeighborhoodDf}
\alias{getNeighborhoodDf}
\title{Create and return the neighborhood data frame for a given time series}
\usage{
tsData.nhoodDf <- getNeighborhoodDf(tsData)
}
\arguments{
\item{tsData}{A time series object.}
}
\value{
A data frame characterizing each time series element and the first difference
}
\description{
A neighborhood of a point p, consists of all the points within a certain distance from p.
For the given time series, the neighborhood data frame consists of each element of
the time series, the previous and next elements, the magnitude of the first difference,
the percent change of the first difference, and the sign of the change.  The column, IS_TP,
is a boolean that indicates whether or not a given element is a transition point.
}
