% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutliers.r
\name{getOutliers}
\alias{getOutliers}
\title{Get the outliers for a given time series}
\usage{
tsData.out <- getOutliers(tsData)
}
\arguments{
\item{tsData}{A time series object.}

\item{outlierType}{A string that determines the way to compute the outliers.  outlierType = c("iqr","stl").  Default is "iqr".}
}
\value{
A vector of indices which represents the location of the outliers within the time series
}
\description{
The outliers are computed using different methods based on the \code{outlierType} parameter
}
\details{
iqr: Find outliers using the interquartile range of the given time series (boxplot method).
stl: Use stl weights.
}
