% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSlopeAngleSize.r
\name{getSlopeAngleSize}
\alias{getSlopeAngleSize}
\title{Compute and return the degree measure of the angle of the slope of the time series.}
\usage{
angleMagnitude <- getSlopeAngleSize(tsData)
}
\arguments{
\item{tsData}{A time series object.}
}
\value{
A numeric value representing the degree measure of the angle of the slope of the time series.
}
\description{
The time series is fitted with a linear model and the degree measure of the angle of the slope is returned.
}
