% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrend.r
\name{getTrend}
\alias{getTrend}
\title{Get the trend of a given time series}
\usage{
tsData.tr <- getTrend(tsData)
}
\arguments{
\item{tsData}{A time series object.}
}
\value{
A time series that represents the trend of the given time series.
}
\description{
This function uses stl to determine a time series' trend
}
