% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrendChangePoints.r
\name{getTrendChangePoints}
\alias{getTrendChangePoints}
\title{Return the change points of the trend for the given time series}
\usage{
tsData.cp <- getTrendChangePoints(tsData, minPctChange, nPeriods)
}
\arguments{
\item{tsData}{A time series object.}

\item{minPctChange}{The optional lower bound a magnitude must exceed}

\item{nPeriods}{An optional number of periods the transition must continue into the future}
}
\value{
A vector of integers corresponding to indices of change points in the given time series trend
}
\description{
Change points are transition points that optionally exceed a certain magnitude and/or
continue in a certain direction for a specified length of time.
}
\details{
Transition points are defined as indices at which the time series changes direction.
We can think of the trend of a univariate time series as analogous to a smooth function
\eqn{f(t)} defined over a given interval of real numbers, \eqn{[a,b]}.  We know that
for \eqn{t in [a,b]}, \eqn{f} is increasing on \eqn{[a,b]} if \eqn{f'(t) > 0},
\eqn{f} is decreasing on \eqn{[a,b]} if \eqn{f'(t) < 0} and
\eqn{f} is constant on \eqn{[a,b]} if \eqn{f'(t) = 0}.
Similarly, we can examine the sign of the difference of a time series at two subsequent
points in time, to determine if the time series is increasing, decreasing, or constant from one
point in time to the next.
Additionally, if we think of the time series as moving through space, a geometric interpretation
would allow us to treat the time series as a vector with a direction (increasing, decreasing, or constant)
and magnitude (the percent change of the time series from one period of time to the next).  This perspective
allows one to adapt what is considered a change point.  Thus, a change point can merely be a transition point,
or it can be a transition point whose percent change from the previous value exceeds some threshold,
or it can be a transition point that causes the time series to continue in the new direction for a
specified period of time.
}
