% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrendOutliers.r
\name{getTrendOutliers}
\alias{getTrendOutliers}
\title{Get the outliers for a given time series}
\usage{
tsData.out <- getTrendOutliers(tsData)
}
\arguments{
\item{tsData}{A time series object.}
}
\value{
A vector of indices which represents the location the outliers within the time series
}
\description{
The outliers are relative to the trend, which is determined by stl
}
