% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrendTransitionPoints.r
\name{getTrendTransitionPoints}
\alias{getTrendTransitionPoints}
\title{Return the transition points of the trend for the given time series}
\usage{
tp <- getTrendTransitionPoints(tsData)
}
\arguments{
\item{tsData}{A time series object.}
}
\value{
A vector of integers corresponding to indices of transition points in the time series
}
\description{
Transition points are defined as indices at which the time series changes direction
}
