% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTsMonth.r
\name{getTsMonth}
\alias{getTsMonth}
\title{Return the two digit month for a each element of a given time series}
\usage{
tsMonth <- getMonth(tsData)
}
\arguments{
\item{tsData}{A time series object.}
}
\value{
A vector of integers (representing months).
}
\description{
The function \code{time} in the \cr\code{stats} package returns the date in fractions.
This function converts that value to the proper integer.
}
\details{
Since yearly (4 digit only) time series have no month and since weeks can span months,
this function is only supported for daily and monthly time series.
}
