% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTsYear.r
\name{getTsYear}
\alias{getTsYear}
\title{Return the four digit year for each element of a given time series}
\usage{
year <- getYear(tsData)
}
\arguments{
\item{tsData}{A Time Series Object.}
}
\value{
A vector of 4 digit integers (representing years).
}
\description{
The function \code{time} in the \cr\code{stats} package returns the date in fractions.
This function converts that value to the proper four digit integer.
}
