% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsToDf.r
\name{tsToDf}
\alias{tsToDf}
\title{Return a monthly or daily univariate time series object as a data frame}
\usage{
tsDf <- tsToDf(tsData)
}
\arguments{
\item{tsData}{A time series object}
}
\value{
A data frame with three or four columns representing year, month number, day of the month number, and the time series value.
}
\description{
Return a monthly or daily univariate time series object as a data frame
}
