% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMA.r
\name{getMA}
\alias{getMA}
\title{Compute a simple moving average for a given time series}
\usage{
tsData.ma <- getMA(tsData,N)
}
\arguments{
\item{tsData}{A time series object.}

\item{N}{An integer indicating how many prior elements to include.}
}
\value{
A time series of moving averages
}
\description{
Compute a moving average using one or more prior elements
}
\details{
If N is null, then 1,2,...,length(tsData) points will be used in the computation of the moving average.
If N is not null, then an N point moving average is computed.
In either case, a one-sided (non-centered) moving average is computed.
The first N-1 elements of the moving average retain the value of the first N-1 elements of the time series.
}
